requirements.txt
vncdotool
pillow